# openapi_chatgpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AudioReader API",
    "description": "API pour generer des audiobooks de haute qualite. Convertit du texte en audio avec synthese vocale avancee, analyse des emotions, dialogues multi-voix et plus.",
    "version": "2.4.0"
  },
  "servers": [
    {
      "url": "https://your-server.com",
      "description": "Serveur de production (remplacer par votre URL)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Serveur local de developpement"
    }
  ],
  "paths": {
    "/voices": {
      "get": {
        "operationId": "listVoices",
        "summary": "Liste les voix disponibles",
        "description": "Retourne la liste des voix disponibles pour la synthese vocale dans la langue specifiee",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "description": "Code langue (fr pour francais, en pour anglais)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "fr",
              "enum": ["fr", "en"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des voix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoicesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/generate": {
      "post": {
        "operationId": "generateAudio",
        "summary": "Genere un fichier audio",
        "description": "Convertit un texte court en fichier audio WAV. Operation asynchrone - utilisez /jobs/{job_id} pour suivre la progression.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generation demarree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/audiobook": {
      "post": {
        "operationId": "generateAudiobook",
        "summary": "Genere un audiobook complet",
        "description": "Genere un audiobook avec analyse automatique des dialogues, attribution des voix aux personnages, analyse des emotions et intonation naturelle. Ideal pour les textes longs comme des chapitres de livres.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudiobookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generation demarree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudiobookResponse"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "getJobStatus",
        "summary": "Statut d'un job",
        "description": "Obtient le statut et la progression d'un job de generation audio",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID du job retourne par /generate ou /audiobook"
          }
        ],
        "responses": {
          "200": {
            "description": "Statut du job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job non trouve"
          }
        }
      }
    },
    "/analyze": {
      "post": {
        "operationId": "analyzeText",
        "summary": "Analyse un texte",
        "description": "Analyse un texte pour detecter les dialogues, identifier les personnages, analyser les emotions et determiner les contours d'intonation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultat de l'analyse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/config": {
      "get": {
        "operationId": "getConfig",
        "summary": "Configuration actuelle",
        "description": "Retourne la configuration actuelle d'AudioReader incluant les fonctionnalites activees et les styles disponibles",
        "responses": {
          "200": {
            "description": "Configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          }
        }
      }
    },
    "/files": {
      "get": {
        "operationId": "listFiles",
        "summary": "Liste les fichiers generes",
        "description": "Retourne la liste des fichiers audio generes avec leurs informations et URLs de telechargement",
        "responses": {
          "200": {
            "description": "Liste des fichiers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/files/{filename}": {
      "get": {
        "operationId": "downloadFile",
        "summary": "Telecharge un fichier",
        "description": "Telecharge un fichier audio genere au format WAV",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nom du fichier a telecharger"
          }
        ],
        "responses": {
          "200": {
            "description": "Fichier audio",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Fichier non trouve"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "VoiceInfo": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "description": "Identifiant unique de la voix"},
          "name": {"type": "string", "description": "Nom de la voix"},
          "gender": {"type": "string", "enum": ["F", "M"], "description": "Genre de la voix"},
          "style": {"type": "string", "description": "Style de la voix (neutral, warm, etc.)"}
        }
      },
      "VoicesResponse": {
        "type": "object",
        "properties": {
          "language": {"type": "string"},
          "kokoro": {"type": "array", "items": {"$ref": "#/components/schemas/VoiceInfo"}},
          "edge_tts": {"type": "array", "items": {"$ref": "#/components/schemas/VoiceInfo"}},
          "recommended": {"type": "string", "description": "Voix recommandee pour cette langue"}
        }
      },
      "GenerateRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {"type": "string", "description": "Texte a convertir en audio"},
          "voice": {"type": "string", "default": "ff_siwis", "description": "ID de la voix"},
          "speed": {"type": "number", "default": 1.0, "minimum": 0.5, "maximum": 2.0, "description": "Vitesse de lecture"},
          "output_name": {"type": "string", "description": "Nom du fichier de sortie (optionnel)"}
        }
      },
      "GenerateResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "job_id": {"type": "string", "description": "ID pour suivre la progression"},
          "status": {"type": "string"},
          "message": {"type": "string"}
        }
      },
      "AudiobookRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {"type": "string", "description": "Texte complet du chapitre ou livre"},
          "title": {"type": "string", "default": "audiobook", "description": "Titre de l'audiobook"},
          "narrator_voice": {"type": "string", "default": "ff_siwis", "description": "Voix du narrateur"},
          "style": {
            "type": "string",
            "enum": ["formal", "conversational", "dramatic", "storytelling", "documentary", "intimate", "energetic"],
            "default": "storytelling",
            "description": "Style de narration"
          },
          "enable_emotions": {"type": "boolean", "default": true, "description": "Activer l'analyse des emotions"},
          "enable_multi_voice": {"type": "boolean", "default": true, "description": "Attribuer des voix differentes aux personnages"},
          "language": {"type": "string", "default": "fr", "enum": ["fr", "en"]}
        }
      },
      "AudiobookResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "job_id": {"type": "string"},
          "status": {"type": "string"},
          "message": {"type": "string"}
        }
      },
      "JobStatusResponse": {
        "type": "object",
        "properties": {
          "job_id": {"type": "string"},
          "status": {"type": "string", "enum": ["pending", "processing", "completed", "failed"]},
          "progress": {"type": "number", "minimum": 0, "maximum": 100, "description": "Progression en pourcentage"},
          "result": {
            "type": "object",
            "properties": {
              "output_file": {"type": "string"},
              "duration_seconds": {"type": "number"},
              "duration_formatted": {"type": "string"},
              "download_url": {"type": "string"},
              "segments_count": {"type": "integer"},
              "characters_detected": {"type": "array", "items": {"type": "string"}}
            }
          },
          "error": {"type": "string"}
        }
      },
      "AnalyzeRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {"type": "string", "description": "Texte a analyser"}
        }
      },
      "AnalyzeResponse": {
        "type": "object",
        "properties": {
          "total_characters": {"type": "integer"},
          "dialogues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {"type": "string"},
                "speaker": {"type": "string"},
                "method": {"type": "string"},
                "confidence": {"type": "number"}
              }
            }
          },
          "characters": {"type": "array", "items": {"type": "string"}},
          "sentences": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {"type": "string"},
                "intonation": {"type": "string"},
                "emotion": {"type": "string"},
                "intensity": {"type": "number"}
              }
            }
          }
        }
      },
      "ConfigResponse": {
        "type": "object",
        "properties": {
          "output_dir": {"type": "string"},
          "default_voice": {"type": "string"},
          "default_language": {"type": "string"},
          "features": {"type": "object"},
          "styles_available": {"type": "array", "items": {"type": "string"}},
          "version": {"type": "string"}
        }
      },
      "FilesResponse": {
        "type": "object",
        "properties": {
          "output_dir": {"type": "string"},
          "files": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {"type": "string"},
                "size_mb": {"type": "number"},
                "modified": {"type": "string"},
                "download_url": {"type": "string"}
              }
            }
          },
          "total_count": {"type": "integer"}
        }
      }
    }
  }
}